refactor liftOver
